+++ lib/chart_it.dart
@@ -6,9 +6,11 @@ export 'src/charts/data/bars/bar_group.dart';
 export 'src/charts/data/bars/bar_series.dart';
 export 'src/charts/data/bars/multi_bar.dart';
 export 'src/charts/data/bars/simple_bar.dart';
-export 'src/charts/data/core/cartesian_data.dart';
-export 'src/charts/data/core/chart_text_style.dart';
-export 'src/charts/data/core/radial_data.dart';
+export 'src/charts/data/core/cartesian/cartesian_data.dart';
+export 'src/charts/data/core/cartesian/cartesian_styling.dart';
+export 'src/charts/data/core/radial/radial_data.dart';
+export 'src/charts/data/core/radial/radial_styling.dart';
+export 'src/charts/data/core/shared/chart_text_style.dart';
 export 'src/charts/data/pie/pie_series.dart';
 export 'src/charts/data/pie/slice_data.dart';
 export 'src/charts/data/pie/slice_data_style.dart';
