+++ lib/src/charts/constants/defaults.dart
@@ -1,4 +1,8 @@
-import 'package:chart_it/chart_it.dart';
+import 'package:chart_it/src/charts/data/bars/bar_data_style.dart';
+import 'package:chart_it/src/charts/data/core/cartesian/cartesian_styling.dart';
+import 'package:chart_it/src/charts/data/core/radial/radial_styling.dart';
+import 'package:chart_it/src/charts/data/core/shared/chart_text_style.dart';
+import 'package:chart_it/src/charts/data/pie/slice_data_style.dart';
 import 'package:flutter/material.dart';
 
 const defaultCartesianChartStyle = CartesianChartStyle(
