+++ lib/src/charts/data/bars/bar_group.dart
@@ -2,8 +2,8 @@ import 'package:chart_it/src/animations/lerps.dart';
 import 'package:chart_it/src/charts/data/bars/bar_data_style.dart';
 import 'package:chart_it/src/charts/data/bars/multi_bar.dart';
 import 'package:chart_it/src/charts/data/bars/simple_bar.dart';
-import 'package:chart_it/src/charts/data/core/cartesian_data.dart';
-import 'package:chart_it/src/charts/data/core/chart_text_style.dart';
+import 'package:chart_it/src/charts/data/core/cartesian/cartesian_data.dart';
+import 'package:chart_it/src/charts/data/core/shared/chart_text_style.dart';
 
 /// Sets the Arrangement for all the bars in a [BarGroup].
 ///
@@ -68,7 +68,7 @@ abstract class BarGroup with ZeroValueProvider<BarGroup> {
     );
   }
 
-  static List<BarGroup> llerp(
+  static List<BarGroup> lerpBarGroupList(
     List<BarGroup>? current,
     List<BarGroup> target,
     double t,
