+++ lib/src/charts/data/core/radial/radial_data.dart
@@ -1,48 +1,11 @@
+import 'package:chart_it/src/charts/data/core/shared/chart_text_style.dart';
 import 'package:equatable/equatable.dart';
 import 'package:flutter/material.dart';
-import 'package:chart_it/src/charts/data/core/chart_text_style.dart';
 
 /// Callback for Mapping a String Value to a Label
 typedef SliceMapper = String Function(num percentage, num value);
 
 abstract class RadialSeries {}
-
-/// Provides the Styling options for any Radial Chart.
-///
-/// Customization Options include BackgroundColor, Start Angle and Axis Styling.
-class RadialChartStyle extends Equatable {
-  /// The Background Color of the Chart.
-  final Color backgroundColor;
-
-  /// The Angle from which the Initial Data Should begin.
-  /// Defaults to zero and starts Vertically Upright.
-  final double initAngle;
-
-  /// Styling for the Axis Lines.
-  final RadialAxis? axis;
-
-  const RadialChartStyle({
-    this.backgroundColor = const Color(0xFFFFECD9),
-    this.initAngle = 0.0,
-    this.axis,
-  }) : assert(
-          initAngle >= 0.0 || initAngle < 360.0,
-          "initAngle should be between 0 to 360 degrees",
-        );
-
-  @override
-  List<Object> get props => [backgroundColor, initAngle];
-
-  RadialChartStyle copyWith({
-    Color? backgroundColor,
-    double? initAngle,
-  }) {
-    return RadialChartStyle(
-      backgroundColor: backgroundColor ?? this.backgroundColor,
-      initAngle: initAngle ?? this.initAngle,
-    );
-  }
-}
 
 /// Defines that the Axis Type is for a [PolarChart], and
 /// provides Styling for Elements only for PolarChart Axis Type
@@ -57,16 +20,14 @@ class PolarAxis extends RadialAxis with EquatableMixin {
 
   @override
   List<Object?> get props => [
-        super.tickStyle,
-        super.gridLineStrokeColor,
-        super.gridLineStrokeWidth,
-      ];
+    super.tickStyle,
+    super.gridLineStrokeColor,
+    super.gridLineStrokeWidth,
+  ];
 
-  PolarAxis copyWith(
-    ChartTextStyle? tickStyle,
-    Color? gridLineStrokeColor,
-    double? gridLineStrokeWidth,
-  ) {
+  PolarAxis copyWith(ChartTextStyle? tickStyle,
+      Color? gridLineStrokeColor,
+      double? gridLineStrokeWidth,) {
     return PolarAxis(
       tickStyle: tickStyle ?? super.tickStyle,
       gridLineStrokeColor: gridLineStrokeColor ?? super.gridLineStrokeColor,
@@ -100,13 +61,13 @@ class RadarAxis extends RadialAxis with EquatableMixin {
 
   @override
   List<Object?> get props => [
-        super.tickStyle,
-        angleLabelStyle,
-        super.gridLineStrokeColor,
-        super.gridLineStrokeWidth,
-        angleLineStrokeColor,
-        angleLineStrokeWidth,
-      ];
+    super.tickStyle,
+    angleLabelStyle,
+    super.gridLineStrokeColor,
+    super.gridLineStrokeWidth,
+    angleLineStrokeColor,
+    angleLineStrokeWidth,
+  ];
 
   RadarAxis copyWith({
     ChartTextStyle? tickStyle,
