+++ lib/src/charts/data/pie/pie_series.dart
@@ -1,10 +1,10 @@
-import 'package:equatable/equatable.dart';
-import 'package:flutter/material.dart';
-import 'package:chart_it/src/charts/data/core/chart_text_style.dart';
-import 'package:chart_it/src/charts/data/core/radial_data.dart';
+import 'package:chart_it/src/charts/data/core/radial/radial_data.dart';
+import 'package:chart_it/src/charts/data/core/shared/chart_text_style.dart';
 import 'package:chart_it/src/charts/data/pie/slice_data.dart';
 import 'package:chart_it/src/charts/data/pie/slice_data_style.dart';
 import 'package:chart_it/src/charts/widgets/pie_chart.dart';
+import 'package:equatable/equatable.dart';
+import 'package:flutter/material.dart';
 
 /// Callback for Label in the Donut Area
 typedef DonutLabel = String Function();
