+++ lib/src/charts/painters/cartesian/cartesian_painter.dart
@@ -1,4 +1,4 @@
-import 'package:chart_it/src/charts/data/core/cartesian_data.dart';
+import 'package:chart_it/src/charts/data/core/cartesian/cartesian_data.dart';
 import 'package:chart_it/src/charts/painters/cartesian/cartesian_chart_painter.dart';
 import 'package:flutter/material.dart';
