+++ lib/src/charts/painters/text/chart_text_painter.dart
@@ -1,5 +1,5 @@
+import 'package:chart_it/src/charts/data/core/shared/chart_text_style.dart';
 import 'package:flutter/cupertino.dart';
-import 'package:chart_it/src/charts/data/core/chart_text_style.dart';
 
 const String _ellipsis = '\u2026';
