+++ lib/src/charts/widgets/core/cartesian_charts.dart
@@ -1,4 +1,5 @@
-import 'package:chart_it/src/charts/data/core/cartesian_data.dart';
+import 'package:chart_it/src/charts/data/core/cartesian/cartesian_data.dart';
+import 'package:chart_it/src/charts/data/core/cartesian/cartesian_styling.dart';
 import 'package:chart_it/src/charts/painters/cartesian/cartesian_chart_painter.dart';
 import 'package:chart_it/src/controllers/cartesian_controller.dart';
 import 'package:flutter/material.dart';
